Refactor admin dashboard components and remove unused files
- Deleted the logoLiceo.png and inscripciones/page.jsx files to streamline the project.
- Refactored loading state handling in the AdminPage and controlSecciones components to utilize a dedicated Loading component for better code organization.
- Updated the CardGridSetion component to improve the display of sections and ensure consistent data handling.
- Enhanced the FormSection and FormSubject components to include better state management and success callbacks.

These changes improve code maintainability, enhance user experience, and remove unnecessary components from the project.

+++ src/context/AuthContext.jsx
@@ -1,23 +1,29 @@
 "use client";
 import { login } from "@/services/login";
+import { logout } from "@/services/logout";
 import { createContext, useContext, useState, useEffect } from "react";
+import { useRouter, usePathname } from "next/navigation";
 
 const AuthContext = createContext();
 
 export function AuthProvider({ children }) {
   const [user, setUser] = useState(null);
   const [loading, setLoading] = useState(true);
+  const router = useRouter();
+  const pathname = usePathname();
+
+  const PUBLIC_ROUTES = ["/", "/register"];
 
   useEffect(() => {
     const storedUser = sessionStorage.getItem("user");
 
     if (storedUser) {
       setUser(JSON.parse(storedUser));
-    } else {
+    } else if (!PUBLIC_ROUTES.includes(pathname)) {
       router.push("/");
     }
     setLoading(false);
-  }, []);
+  }, [pathname]);
 
   /**
    * Funcion para iniciar sesión
@@ -42,13 +48,17 @@ export function AuthProvider({ children }) {
    * @returns {void}
    */
   const handleLogout = async () => {
-    sessionStorage.clear();
-    setUser(null);
-    router.push("/");
+    const result = await logout();
+
+    if (result.ok) {
+      sessionStorage.clear();
+      setUser(null);
+      router.push("/");
+    }
   };
 
   return (
-    <AuthContext.Provider value={{ user, loading }}>
+    <AuthContext.Provider value={{ user, loading, handleLogout, handleLogin }}>
       {children}
     </AuthContext.Provider>
   );
